requirements.txt
discord.py==2.3.2
fuzzywuzzy==0.18.0
toml==0.10.2
python-Levenshtein==0.25.0
jishaku==2.5.2
tortoise-orm==0.20.0
